+++ Lib/TSCompiler/DepedencyUpdate.cs
@@ -1,0 +1,9 @@
+﻿namespace Lib.TSCompiler
+{
+    public enum DepedencyUpdate
+    {
+        Disabled,
+        Install,
+        Upgrade
+    }
+}