+++ Lib/TSCompiler/StyleDefNamingStyle.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text.RegularExpressions;
+using Lib.DiskCache;
+using Lib.Utils;
+
+namespace Lib.TSCompiler
+{
+    public enum StyleDefNamingStyle
+    {
+        AddNames,
+        PreserveNames,
+        RemoveNames
+    }
+}